extensions/time: Need to require json/add/core
The reason for this is that we overwrite an extension which is also provided in json/add/core. If we do _not_ load this early on, there is a slight risk that the Time extension gets loaded prematurely, breaking the functionality. I just ran into this in a project, which is why I'm fixing it here as well.