Added Hashable mixin
This is part of a larger refactoring, where this class and a bunch of other classes are taken out of uxFactory Server and moved to other gems in the ecosystem.

In conjuction with this, the `Hashable` mixin is also moved to a namespace. It has previously existed as a top-level module, and it will be aliased as such for legacy scenarios where we need to retain compatibility with code written/generated that way.